fix: ensure repositories are correctly marked with `inherited` creds in CLI output (#13428)
* tests: ensure `InheritedCreds` is propagated via repo API endpoints

Signed-off-by: OneMatchFox <878612+onematchfox@users.noreply.github.com>

* fix: ensure `InheritedCreds` is propagated via repo API endpoints

Signed-off-by: OneMatchFox <878612+onematchfox@users.noreply.github.com>

* tests: add e2e test for `argocd repo get` with inherited credentials

Signed-off-by: OneMatchFox <878612+onematchfox@users.noreply.github.com>

* fix(cli): prioritise value of `InheritedCreds` over `HasCredentials()`

Since the API does not return sensitive information `HasCredentials()` will return false for all scenarios except when username/password is used as credentials. Given the current logic this means that the code will never even check `InheritedCreds` resulting in an output of `false` for `CREDS` column (in the case of inherited credentials).

Note: There remains a bug in this code in that any repo that has explicit (sensitive) credentials (e.g. SSH private key) will still be displayed as `CREDS = false`.
Signed-off-by: OneMatchFox <878612+onematchfox@users.noreply.github.com>

---------

Signed-off-by: OneMatchFox <878612+onematchfox@users.noreply.github.com>